feat feature management offline on mobile

+++ front/src/components/WorkingTimes/WorkingTimesTable.vue
@@ -54,57 +54,55 @@
 </template>
 
 <script>
-import Cookies from 'js-cookie';
-import jwt_decode from 'jwt-decode';
-import Card from '../Cards/Card';
-import LTable from 'src/components/Table';
-import AddWorkingTime from './AddWorkingTime';
+import Cookies from "js-cookie";
+import jwt_decode from "jwt-decode";
+import Card from "../Cards/Card";
+import LTable from "src/components/Table";
+import AddWorkingTime from "./AddWorkingTime";
 import {
   getAllWorkingTimesByUserId,
-  deleteWorkingtimeById,
-} from '../../api_wrapper/workingtimes/workingtime';
+  deleteWorkingtimeById
+} from "../../api_wrapper/workingtimes/workingtime";
+import { getDeviceType } from "../../helpers/getDeviceType";
+import { getConnexionType } from "../../helpers/getConnexionType";
 
 export default {
-  name: 'WorkingTimesTable',
+  name: "WorkingTimesTable",
   components: {
     Card,
     LTable,
-    AddWorkingTime,
+    AddWorkingTime
   },
   props: {
     mini: {
       type: Boolean,
-      required: true,
+      required: true
     },
-    selectedUserId: String,
+    selectedUserId: String
   },
   computed: {
     isMobile() {
-      if (screen.width <= 760) {
-        return true;
-      } else {
-        return false;
-      }
-    },
+      return getDeviceType() === "mobile";
+    }
   },
   data() {
     return {
-      userId: '',
+      userId: "",
       workingTimes: [],
-      tableColumns: ['Start', 'End', 'Actions'],
+      tableColumns: ["Start", "End", "Actions"],
       showModal: false,
-      selectedId: String,
+      selectedId: String
     };
   },
 
   methods: {
     formatDates(date) {
-      return `${date.toLocaleDateString('fr-FR')} at ${date.toLocaleTimeString(
-        'fr-FR',
+      return `${date.toLocaleDateString("fr-FR")} at ${date.toLocaleTimeString(
+        "fr-FR",
         {
-          hour: '2-digit',
-          minute: '2-digit',
-        },
+          hour: "2-digit",
+          minute: "2-digit"
+        }
       )}`;
     },
 
@@ -116,34 +114,49 @@ export default {
       const workingTimes = await getAllWorkingTimesByUserId(
         this.userId,
         now.toISOString(),
-        endDate.toISOString(),
+        endDate.toISOString()
       );
-      const workingtimetable = [];
-      for (const time in workingTimes) {
-        const dateStart = new Date(workingTimes[time].start);
-        const dateEnd = new Date(workingTimes[time].end);
-        if (dateStart < now) {
-          workingtimetable.push(time);
-        } else {
-          workingTimes[time].start = this.formatDates(dateStart);
-          workingTimes[time].end = this.formatDates(dateEnd);
+      if (!!workingTimes) {
+        const workingtimetable = [];
+        for (const time in workingTimes) {
+          const dateStart = new Date(workingTimes[time].start);
+          alert(workingTimes[time].start);
+          const dateEnd = new Date(workingTimes[time].end);
+          if (dateStart < now) {
+            workingtimetable.push(time);
+          } else {
+            workingTimes[time].start = this.formatDates(dateStart);
+            workingTimes[time].end = this.formatDates(dateEnd);
+          }
+        }
+        for (const element in workingtimetable) {
+          workingTimes.splice(element, 1);
+        }
+        if (this.mini == true) {
+          workingTimes.splice(3);
         }
       }
-      for (const element in workingtimetable) {
-        workingTimes.splice(element, 1);
-      }
-      if (this.mini == true) {
-        workingTimes.splice(3);
-      }
-      return workingTimes;
+      return [];
     },
     handleEdit(id) {
+      if (getConnexionType() === "none") {
+        alert(
+          "Oops ! You must be connected to the Internet to use this feature"
+        );
+        return;
+      }
       this.selectedId = id.toString();
       this.showModal = true;
       window.scrollTo(0, 0);
     },
     handleDelete(workingtimeId) {
-      if (confirm('Do you really want to delete this working time ?')) {
+      if (getConnexionType() === "none") {
+        alert(
+          "Oops ! You must be connected to the Internet to use this feature"
+        );
+        return;
+      }
+      if (confirm("Do you really want to delete this working time ?")) {
         this.deleteWorkingTime(workingtimeId);
       }
     },
@@ -153,21 +166,22 @@ export default {
     },
     async affectWorkingTimes() {
       this.workingTimes = await this.getWorkingTimes();
+      alert(this.workingTimes[0]);
     },
     switchModal() {
       this.showModal = !this.showModal;
-    },
+    }
   },
 
   mounted() {
     if (!this.selectedUserId) {
-      const token = Cookies.get('token');
+      const token = Cookies.get("token");
       this.userId = jwt_decode(token).id;
     } else {
       this.userId = this.selectedUserId;
     }
     this.affectWorkingTimes();
-  },
+  }
 };
 </script>
